Initial more aggressive position refactor
Update test fixtures

Generalizes edit new action

Rename finalStages to actionStages

More tweaks

Refactor EditNew into multiple files
