+++ src/processTargets/processTargets.ts
@@ -194,13 +194,29 @@ function processPrimitiveTarget(
   const markOutputTargets = markStage.run(context);
 
   /**
+   * The modifiers from the target descriptor. Note that we reverse them before
+   * running them through the pipeline because they arrive in the same order in
+   * which they're spoken.
+   */
+  const modifiers = [...targetDescriptor.modifiers].reverse();
+
+  /**
+   * If the pipeline ends in a position modifier, we run in after the action
+   * stages
+   */
+  const [initialModifiers, finalModifiers] =
+    modifiers.at(-1)?.type === "position"
+      ? [modifiers.slice(0, -1), [modifiers.at(-1)!]]
+      : [modifiers, []];
+
+  /**
    * The modifier pipeline that will be applied to construct our final targets
    */
   const modifierStages = [
     // Reverse target modifiers because they are returned in reverse order from
     // the api, to match the order in which they are spoken.
     ...targetDescriptor.modifiers.map(getModifierStage).reverse(),
-    ...context.finalStages,
+    ...context.actionStages,
   ];
 
   /**
