+++ src/processTargets/targets/index.ts
@@ -1,0 +1,26 @@
+export * from "./BaseTarget";
+export { default as BaseTarget } from "./BaseTarget";
+export * from "./DocumentTarget";
+export { default as DocumentTarget } from "./DocumentTarget";
+export * from "./LineTarget";
+export { default as LineTarget } from "./LineTarget";
+export * from "./NotebookCellTarget";
+export { default as NotebookCellTarget } from "./NotebookCellTarget";
+export * from "./ParagraphTarget";
+export { default as ParagraphTarget } from "./ParagraphTarget";
+export * from "./PlainTarget";
+export { default as PlainTarget } from "./PlainTarget";
+export * from "./PositionTarget";
+export { default as PositionTarget } from "./PositionTarget";
+export * from "./RawSelectionTarget";
+export { default as RawSelectionTarget } from "./RawSelectionTarget";
+export * from "./ScopeTypeTarget";
+export { default as ScopeTypeTarget } from "./ScopeTypeTarget";
+export * from "./SubTokenWordTarget";
+export { default as SubTokenWordTarget } from "./SubTokenWordTarget";
+export * from "./SurroundingPairTarget";
+export { default as SurroundingPairTarget } from "./SurroundingPairTarget";
+export * from "./TokenTarget";
+export { default as TokenTarget } from "./TokenTarget";
+export * from "./WeakTarget";
+export { default as WeakTarget } from "./WeakTarget";
