+++ src/test/suite/fixtures/recorded/hatTokenMap/bringHarpAndPointToEndOfThisAndEndOfWhaleTakeWhale.yml
@@ -54,6 +54,8 @@ finalState:
   thatMark:
     - anchor: {line: 0, character: 12}
       active: {line: 0, character: 18}
+    - anchor: {line: 0, character: 12}
+      active: {line: 0, character: 18}
   sourceMark:
     - anchor: {line: 0, character: 0}
       active: {line: 0, character: 5}
