+++ src/test/suite/fixtures/recorded/hatTokenMap/bringPointAndHarpToEndOfThisAndEndOfWhaleTakeWhale.yml
@@ -50,8 +50,10 @@ finalState:
       end: {line: 0, character: 5}
     default.w:
       start: {line: 0, character: 7}
-      end: {line: 0, character: 12}
+      end: {line: 0, character: 18}
   thatMark:
+    - anchor: {line: 0, character: 12}
+      active: {line: 0, character: 18}
     - anchor: {line: 0, character: 12}
       active: {line: 0, character: 18}
   sourceMark:
