+++ src/test/suite/fixtures/recorded/positions/bringHarpAfterLineOatsAndWhale.yml
@@ -1,0 +1,56 @@
+languageId: typescript
+command:
+  spokenForm: bring harp after line oats and whale
+  version: 2
+  targets:
+    - type: primitive
+      mark: {type: decoratedSymbol, symbolColor: default, character: h}
+    - type: list
+      elements:
+        - type: primitive
+          mark: {type: decoratedSymbol, symbolColor: default, character: o}
+          modifiers:
+            - {type: position, position: after}
+            - type: containingScope
+              scopeType: {type: line}
+        - type: primitive
+          mark: {type: decoratedSymbol, symbolColor: default, character: w}
+  usePrePhraseSnapshot: true
+  action: {name: replaceWithTarget}
+initialState:
+  documentContents: |-
+    const whatever = "hello";
+
+    const whatever = "hello";
+  selections:
+    - anchor: {line: 2, character: 24}
+      active: {line: 2, character: 24}
+  marks:
+    default.h:
+      start: {line: 0, character: 6}
+      end: {line: 0, character: 14}
+    default.o:
+      start: {line: 0, character: 18}
+      end: {line: 0, character: 23}
+    default.w:
+      start: {line: 2, character: 6}
+      end: {line: 2, character: 14}
+finalState:
+  documentContents: |-
+    const whatever = "hello";
+    whatever
+
+    const whatever = "hello";
+    whatever
+  selections:
+    - anchor: {line: 3, character: 24}
+      active: {line: 3, character: 24}
+  thatMark:
+    - anchor: {line: 1, character: 0}
+      active: {line: 1, character: 8}
+    - anchor: {line: 4, character: 0}
+      active: {line: 4, character: 8}
+  sourceMark:
+    - anchor: {line: 0, character: 6}
+      active: {line: 0, character: 14}
+fullTargets: [{type: primitive, mark: {type: decoratedSymbol, symbolColor: default, character: h}, modifiers: []}, {type: list, elements: [{type: primitive, mark: {type: decoratedSymbol, symbolColor: default, character: o}, modifiers: &ref_0 [{type: position, position: after}, {type: containingScope, scopeType: {type: line}}]}, {type: primitive, mark: {type: decoratedSymbol, symbolColor: default, character: w}, modifiers: *ref_0}]}]
