+++ src/typings/Types.ts
@@ -24,7 +24,11 @@ export interface Token extends FullRangeInfo {
 }
 
 export interface ProcessedTargetsContext {
-  finalStages: ModifierStage[];
+  /**
+   * Modifier stages contributed by the action. These stages will run at the
+   * end of the pipeline, but before the final positional stage, if there is one
+   */
+  actionStages: ModifierStage[];
   currentSelections: SelectionWithEditor[];
   currentEditor: vscode.TextEditor | undefined;
   hatTokenMap: ReadOnlyHatMap;
